updated loading visuals
