+++ src/components/chat/chat-message-skeleton.tsx
@@ -1,8 +1,7 @@
 'use client';
 
-import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
-import { Bot } from 'lucide-react';
 import { LoaderFive } from '../ui/loader';
+import { getPersonaColor } from '@/lib/persona-colors';
 
 interface Persona {
     id: string;
@@ -15,17 +14,20 @@ interface ChatMessageSkeletonProps {
 }
 
 export function ChatMessageSkeleton({ persona }: ChatMessageSkeletonProps) {
+    const personaColorClass = getPersonaColor(persona?.id);
+
     return (
-        <div className="group flex items-start gap-3">
-            <Avatar className="h-8 w-8 bg-accent/50 text-accent-foreground border border-accent">
-                <AvatarImage src={persona?.avatarUrl} alt={persona?.name} />
-                <AvatarFallback className="bg-transparent">
-                    <Bot className="h-5 w-5" />
-                </AvatarFallback>
-            </Avatar>
-            <div className="flex flex-col gap-1 items-start">
-                <div className="max-w-2xl p-3 px-4 text-base rounded-2xl bg-secondary">
-                    <div className="flex items-center gap-2 text-muted-foreground">
+        <div className="group flex w-full gap-3 px-4 py-3 justify-start">
+            {/* spacing where avatar would be */}
+            <div className="h-8 w-8" />
+            <div className="flex flex-col gap-1 max-w-2xl items-start">
+                {persona?.name && (
+                    <div className="flex items-center gap-1 px-1">
+                        <p className="text-xs font-medium text-foreground/60">{persona.name}</p>
+                    </div>
+                )}
+                <div className={`relative text-sm leading-relaxed text-foreground/90 border-l-2 pl-4 pr-3 py-3 ${personaColorClass}`}>
+                    <div className="font-sans font-bold text-muted-foreground">
                         <LoaderFive text={`${persona?.name || 'AI Assistant'} is thinking...`} />
                     </div>
                 </div>
